refactor(explorer/player): Simplify mock util function imports
Export graphql type mocking functions direction from main namespace; deprecate utils and queries
exports as utils are internal to graphql-sdk and queries should be application-specific

BREAKING CHANGE: Updated graphql-sdk module exports; deprecated utils and queries, mocking functions
are now top-level exports

re #110
