+++ packages/explorer/src/views/Account/enhance.js
@@ -3,15 +3,12 @@ import { compose, withHandlers } from 'recompose'
 import { graphql } from 'react-apollo'
 import gql from 'graphql-tag'
 import { MathBN } from '../../utils'
-import { utils as gqlSDKUtils } from '@livepeer/graphql-sdk'
+import { mockAccount } from '@livepeer/graphql-sdk'
 import {
   connectCoinbaseQuery,
   connectCurrentRoundQuery,
   connectToasts,
-  withTransactionHandlers,
 } from '../../enhancers'
-
-const { mockAccount } = gqlSDKUtils
 
 const accountQuery = gql`
   fragment AccountFragment on Account {
@@ -188,6 +185,5 @@ export default compose(
   connectCurrentRoundQuery,
   connectMeDelegatorTranscoderQuery,
   connectToasts,
-  withTransactionHandlers,
   mapMutationHandlers,
 )
