+++ packages/explorer/src/views/AccountBroadcasting/enhance.js
@@ -1,10 +1,8 @@
 import { compose } from 'recompose'
 import { graphql } from 'react-apollo'
 import gql from 'graphql-tag'
-import { utils as gqlSDKUtils } from '@livepeer/graphql-sdk'
+import { mockBroadcaster } from '@livepeer/graphql-sdk'
 import { connectCoinbaseQuery, withTransactionHandlers } from '../../enhancers'
-
-const { mockBroadcaster } = gqlSDKUtils
 
 const AccountBroadcasterAndJobsQuery = gql`
   fragment BroadcasterFragment on Broadcaster {
