+++ packages/explorer/src/views/AccountDelegating/enhance.js
@@ -1,15 +1,13 @@
 import { compose, lifecycle, mapProps, withHandlers } from 'recompose'
 import { graphql, withApollo } from 'react-apollo'
 import gql from 'graphql-tag'
-import { utils as gqlSDKUtils } from '@livepeer/graphql-sdk'
+import { mockDelegator } from '@livepeer/graphql-sdk'
 import {
   connectCoinbaseQuery,
   connectCurrentRoundQuery,
   connectToasts,
   withTransactionHandlers,
 } from '../../enhancers'
-
-const { mockDelegator } = gqlSDKUtils
 
 const AccountDelegatorQuery = gql`
   fragment DelegatorFragment on Delegator {
